requirements.txt
pandas
psycopg2-binary
sqlalchemy
dbt-core
dbt-postgres
python-dotenv
astronomer-cosmos